Refactor attack table editor to own file for reuse
