+++ src/components/inputs/index.ts
@@ -2,3 +2,4 @@ export * from './CheckboxGroup';
 export * from './CheckboxInput';
 export * from './LabeledInput';
 export * from './LabeledSelect';
+export * from './AttackTableEditor';
